Fix error handling
We had a custom "addError" Method on the parser context, that was writing string errors into an array and is never queried except in one unit tests. I replaced it with the LoggerInterface. In doing so I had to create the Gridtablebuilder with DI as well, it had been created with a constructor until now.

In doing so I also found 4 failing functional tests where  errors have been swallowed until now and will address them next